hibernate_repo_check_writer_not_null

+++ src/main/java/edu/alenkin/repository/hibernate/WriterRepositoryHiber.java
@@ -1,6 +1,5 @@
 package edu.alenkin.repository.hibernate;
 
-import edu.alenkin.model.Post;
 import edu.alenkin.model.Writer;
 import edu.alenkin.repository.WriterRepository;
 import edu.alenkin.utils.hiber.HibernateWorker;
@@ -23,24 +22,26 @@ public class WriterRepositoryHiber implements WriterRepository {
 
     @Override
     public void delete(Long writerId) {
-        try (Session session = HibernateWorker.getSessionFactory().openSession()) {
+        try (Session session = HibernateWorker.getSession()) {
             session.beginTransaction();
             Writer currentWriter = session.get(Writer.class, writerId);
-            session.delete(currentWriter);
+            if (currentWriter != null) {
+                session.delete(currentWriter);
+            }
             session.getTransaction().commit();
         }
     }
 
     @Override
     public Writer get(Long writerId) {
-        try (Session session = HibernateWorker.getSessionFactory().openSession()) {
+        try (Session session = HibernateWorker.getSession()) {
             return session.get(Writer.class, writerId);
         }
     }
 
     @Override
     public void delete(Writer writer) {
-        try (Session session = HibernateWorker.getSessionFactory().openSession()) {
+        try (Session session = HibernateWorker.getSession()) {
             session.beginTransaction();
             session.delete(writer);
             session.getTransaction().commit();
@@ -49,7 +50,7 @@ public class WriterRepositoryHiber implements WriterRepository {
 
     @Override
     public Long save(Writer writer) {
-        try (Session session = HibernateWorker.getSessionFactory().openSession()) {
+        try (Session session = HibernateWorker.getSession()) {
             session.beginTransaction();
             Long id = null;
             if (writer.getId() != null) {
@@ -64,7 +65,7 @@ public class WriterRepositoryHiber implements WriterRepository {
 
     @Override
     public List<Writer> getAll() {
-        try (Session session = HibernateWorker.getSessionFactory().openSession()) {
+        try (Session session = HibernateWorker.getSession()) {
             Query query = session.createQuery("from Writer");
             List<Writer> writers = query.getResultList();
             writers = (writers == null || writers.isEmpty())
